chore(refactor): use LayoutEngine in CanvasBuilder
