+++ web_src/src/lib/ai/canvas-builder.ts
@@ -1,11 +1,14 @@
 import type {
+  BlueprintsBlueprint,
   CanvasesCanvas,
+  ComponentsComponent,
   ComponentsEdge,
   ComponentsIntegrationRef,
   ComponentsNode,
   ComponentsNodeType,
   OrganizationsIntegration,
 } from "@/api-client";
+import type { LayoutEngine } from "@/lib/layout";
 import type { BuildingBlock, BuildingBlockCategory } from "@/ui/BuildingBlocksSidebar";
 import { generateNodeId, generateUniqueNodeName } from "@/pages/workflowv2/utils";
 import type {
@@ -24,41 +27,51 @@ type ScoredChannel = {
   description: string;
 };
 
+export type CanvasBuilderOptions = {
+  canvas: CanvasesCanvas;
+  buildingBlocks: BuildingBlockCategory[];
+  integrations?: OrganizationsIntegration[];
+  components?: ComponentsComponent[];
+  blueprints?: BlueprintsBlueprint[];
+};
+
 /*
  * Applies a list of AI-generated canvas operations to a canvas.
  * The operations are applied in the order they are provided.
  *
- * @param canvas - The canvas to apply the operations to.
- * @param buildingBlocks - The building blocks to use for the canvas.
- * @param integrations - The integrations to use for the canvas.
+ * @param options - Builder configuration with canvas, building blocks and optional metadata.
  * @returns The updated canvas.
  */
 export class CanvasBuilder {
   private readonly canvas: CanvasesCanvas;
   private readonly integrations: OrganizationsIntegration[];
   private readonly blockLookup: Map<string, BuildingBlockCategory["blocks"][number]>;
+  private readonly layoutComponents: ComponentsComponent[];
+  private readonly layoutBlueprints: BlueprintsBlueprint[];
   private createdNodeIdsByKey: Map<string, string>;
   private addedNodeBlockNameByKey: Map<string, string>;
+  private addedNodeIds: Set<string>;
   private updatedNodes: ComponentsNode[];
   private updatedEdges: ComponentsEdge[];
 
-  constructor(
-    canvas: CanvasesCanvas,
-    buildingBlocks: BuildingBlockCategory[],
-    integrations: OrganizationsIntegration[] = [],
-  ) {
+  constructor(options: CanvasBuilderOptions) {
+    const { canvas, buildingBlocks, integrations = [], components = [], blueprints = [] } = options;
+
     this.canvas = canvas;
     this.integrations = integrations;
     this.blockLookup = new Map(
       buildingBlocks.flatMap((category) => category.blocks.map((block) => [block.name, block])),
     );
+    this.layoutComponents = components;
+    this.layoutBlueprints = blueprints;
     this.createdNodeIdsByKey = new Map();
     this.addedNodeBlockNameByKey = new Map();
+    this.addedNodeIds = new Set();
     this.updatedNodes = [];
     this.updatedEdges = [];
   }
 
-  apply(operations: CanvasOperation[]): CanvasesCanvas {
+  async apply(operations: CanvasOperation[], layout?: LayoutEngine): Promise<CanvasesCanvas> {
     this.initializeState(operations);
 
     for (const operation of operations) {
@@ -79,7 +92,7 @@ export class CanvasBuilder {
       }
     }
 
-    return {
+    const updatedCanvas = {
       ...this.canvas,
       spec: {
         ...this.canvas.spec,
@@ -87,11 +100,14 @@ export class CanvasBuilder {
         edges: this.updatedEdges,
       },
     };
+
+    return this.applyLayoutIfNewNodesWereAdded(updatedCanvas, layout);
   }
 
   private initializeState(operations: CanvasOperation[]): void {
     this.createdNodeIdsByKey = new Map();
     this.addedNodeBlockNameByKey = new Map();
+    this.addedNodeIds = new Set();
     this.updatedNodes = [...(this.canvas.spec?.nodes || [])];
     this.updatedEdges = [...(this.canvas.spec?.edges || [])];
 
@@ -325,6 +341,7 @@ export class CanvasBuilder {
     }
 
     this.updatedNodes.push(newNode);
+    this.addedNodeIds.add(newNodeId);
     if (operation.nodeKey) {
       this.createdNodeIdsByKey.set(operation.nodeKey, newNodeId);
     }
@@ -411,13 +428,35 @@ export class CanvasBuilder {
       return;
     }
 
+    const targetNodeId = this.updatedNodes[nodeIndex]?.id;
     this.updatedNodes.splice(nodeIndex, 1);
+    if (typeof targetNodeId === "string") {
+      this.addedNodeIds.delete(targetNodeId);
+    }
     for (let edgeIndex = this.updatedEdges.length - 1; edgeIndex >= 0; edgeIndex -= 1) {
       const edge = this.updatedEdges[edgeIndex];
       if (edge.sourceId === targetId || edge.targetId === targetId) {
         this.updatedEdges.splice(edgeIndex, 1);
       }
     }
+  }
+
+  private async applyLayoutIfNewNodesWereAdded(
+    updatedCanvas: CanvasesCanvas,
+    layout?: LayoutEngine,
+  ): Promise<CanvasesCanvas> {
+    if (!layout) {
+      return updatedCanvas;
+    }
+
+    if (this.addedNodeIds.size === 0) {
+      return updatedCanvas;
+    }
+
+    return layout.apply(updatedCanvas, {
+      components: this.layoutComponents,
+      blueprints: this.layoutBlueprints,
+    });
   }
 }
