+++ web_src/src/lib/layout/index.ts
@@ -1,0 +1,5 @@
+export * from "./types";
+export * from "./elk";
+import { ElkLayoutEngine } from "./elk";
+
+export const DefaultLayoutEngine = new ElkLayoutEngine();
